@@ -4,7 +4,7 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "start": "nx serve",
     "build": "nx build",
     "test": "nx test",
@@ -12,53 +12,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
-    "@nrwl/angular": "14.1.2",
+    "@angular/animations": "14.0.0",
+    "@angular/common": "14.0.0",
+    "@angular/compiler": "14.0.0",
+    "@angular/core": "14.0.0",
+    "@angular/forms": "14.0.0",
+    "@angular/platform-browser": "14.0.0",
+    "@angular/platform-browser-dynamic": "14.0.0",
+    "@angular/router": "14.0.0",
+    "@nrwl/angular": "14.2.0-rc.2",
     "rxjs": "~7.4.0",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.0",
-    "@angular-eslint/eslint-plugin": "~13.1.0",
-    "@angular-eslint/eslint-plugin-template": "~13.1.0",
-    "@angular-eslint/template-parser": "~13.1.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "~13.3.0",
-    "@angular/language-service": "~13.3.0",
+    "@angular-devkit/build-angular": "14.0.0",
+    "@angular-eslint/eslint-plugin": "13.2.1",
+    "@angular-eslint/eslint-plugin-template": "13.2.1",
+    "@angular-eslint/template-parser": "13.2.1",
+    "@angular/cli": "~14.0.0",
+    "@angular/compiler-cli": "14.0.0",
+    "@angular/language-service": "14.0.0",
     "@nrwl/cli": "14.0.3",
-    "@nrwl/cypress": "14.1.2",
-    "@nrwl/eslint-plugin-nx": "14.1.2",
-    "@nrwl/jest": "14.1.2",
-    "@nrwl/linter": "14.1.2",
-    "@nrwl/nx-cloud": "14.0.3",
-    "@nrwl/workspace": "14.1.2",
+    "@nrwl/cypress": "14.2.0-rc.2",
+    "@nrwl/eslint-plugin-nx": "14.2.0-rc.2",
+    "@nrwl/jest": "14.2.0-rc.2",
+    "@nrwl/linter": "14.2.0-rc.2",
+    "@nrwl/nx-cloud": "14.0.5",
+    "@nrwl/workspace": "14.2.0-rc.2",
     "@types/jest": "27.4.1",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "~5.18.0",
-    "@typescript-eslint/parser": "~5.18.0",
+    "@typescript-eslint/eslint-plugin": "5.24.0",
+    "@typescript-eslint/parser": "5.24.0",
     "cypress": "^9.1.0",
-    "eslint": "~8.12.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
     "jest": "27.5.1",
-    "jest-preset-angular": "11.1.1",
-    "ng-packagr": "^13.0.0",
-    "nx": "14.1.2",
+    "jest-preset-angular": "11.1.2",
+    "ng-packagr": "14.0.0",
+    "nx": "14.2.0-rc.2",
     "postcss": "^8.4.5",
-    "postcss-import": "^14.0.2",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^10.1.1",
-    "prettier": "^2.5.1",
+    "postcss-import": "14.1.0",
+    "postcss-preset-env": "7.5.0",
+    "postcss-url": "10.1.3",
+    "prettier": "2.6.2",
     "ts-jest": "27.1.4",
-    "ts-node": "9.1.1",
-    "typescript": "~4.6.2"
+    "ts-node": "10.8.1",
+    "typescript": "4.7.3"
   }
 }
