@@ -1,6 +1,6 @@
-module.exports = {
+export default {
   displayName: 'cart',
-  preset: '../../jest.preset.ts',
+  preset: '../../jest.preset.js',
   setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
   globals: {
     'ts-jest': {
